Omnifying: add instructions to create vendor/$VENDOR/$DEVICE/Android.mk
Omni rom needs instructions for the libqcci_legacy module. Unfortunately,
this module is proprietary. The module is currently copied out during
extract-files, and it is included to be copied back into the rom. However,
some packages need to know what symbols libqcci_legacy exports, so we need
a place to define the module, similar to the modules defined in
vendor/lge/g3-common/Android.mk, which is also created by this file. These
instructions tell make where to export the symbols so other packages can
make reference to them.